Move InfoPane to subfolder
